refactor: move files in db folder
