+++ src/db/generics/generic.entity.ts
